NGFW-10268 Move alert processing.
